+++ composeApp/src/jvmMain/kotlin/ai/mcpdirect/mcpdirectstudioapp/App.kt
@@ -18,6 +18,7 @@ import ai.mcpdirect.studio.app.mcpkey.MCPAccessKeyScreen
 import ai.mcpdirect.studio.app.setting.SettingsScreen
 import ai.mcpdirect.studio.app.team.MCPTeamScreen
 import ai.mcpdirect.studio.app.team.MCPTeamToolMakerScreen
+import ai.mcpdirect.studio.app.team.MCPTeamToolMakerTemplateScreen
 import ai.mcpdirect.studio.app.theme.purple.PurpleTheme
 import ai.mcpdirect.studio.app.tool.MCPToolsScreen
 import ai.mcpdirect.studio.app.tool.ToolDetailScreen
@@ -82,6 +83,9 @@ fun App(){
                         }
                         Screen.MCPTeamToolMaker -> {
                             MCPTeamToolMakerScreen()
+                        }
+                        Screen.MCPTeamToolMakerTemplate -> {
+                            MCPTeamToolMakerTemplateScreen()
                         }
                         Screen.VirtualMCP -> {
                             VirtualMakerScreen()